requirements.txt
discord.py==2.3.2
openai==0.27.8
Pillow==10.0.0
python-dotenv==1.0.0
Requests==2.31.0
urllib3==2.0.3